+++ Car_rental_PAI_project/src/main/java/com/example/Car_rental_PAI_project/service/ClientService.java
@@ -5,7 +5,7 @@ import com.example.Car_rental_PAI_project.model.Client;
 import java.util.Optional;
 
 public interface ClientService {
-    Optional<Client> getClient(Long client_Id);
+    Optional<Client> getClient(Integer client_Id);
     Client setClient(Client client);
-    void deleteClient(Long client_Id);
+    void deleteClient(Integer client_Id);
 }
