+++ Car_rental_PAI_project/src/main/java/com/example/Car_rental_PAI_project/service/DepartmentService.java
@@ -5,7 +5,7 @@ import com.example.Car_rental_PAI_project.model.Department;
 import java.util.Optional;
 
 public interface DepartmentService {
-    Optional<Department> getDepartment(Long department_Id);
+    Optional<Department> getDepartment(Integer department_Id);
     Department setDepartment(Department department);
-    void deleteDepartment(Long department_Id);
+    void deleteDepartment(Integer department_Id);
 }
